refactor(pino-http-send): separate provider implementations into modules
Provider-specific logic was intertwined with core transport logic in a
single file, making the code harder to maintain and harder to extend with
new providers. Extracted exception handler and OpenObserve providers into
separate modules with shared utilities, improving maintainability and
enabling cleaner addition of future providers.
